Adding new class for non-keri event SADs.
Allowing SAID generation to ignore certain fields.

Database and REST API for new controller Notifications.

Upgrade to signing data at rest to allow for a more comprehensive and secure solution.

Signaling infrastructure for non-durable signals to the controller of the agent (to reload data, for example).

Integration of new notifier and signaler APIs into agent handlers.

Removed KLI pytests that are brittle, slow and cumbersome to debug.  Excluded the cli directory from coverage for now until we come up with an alternate solution to tests

Signed-off-by: pfeairheller <pfeairheller@gmail.com>