fix: respect headDirPath from --config-dir
The headDirPath for both the Regery and the ` is not set from the

The --config argument to init.py was never used so I added the "-dir" suffix to clarify the intent of the argument.

The Keeper did not recognize the headDirPath passed in through the Configer instance in ****kwargs['cf']**. This was causing the Keeper to create files in either the default HeadDirPath or the AltHeadDirPath when everything else was being created in the configured `headDirPath`.