feat: unify CLI args for incept and rotate
This adds in the support for using a configuration file (JSON) to configure a rotation event as well as support for using command line arguments to configure an inception event for all relevant configuration properties of each type of event.

The --file argument is no longer required for "incept" nor is it required for "rotate" though there are a minimum set of required attributes checked for "incept" that maps to the existing sample configuration "transferable-sample.json" file.

There are a few code cleanups like the `config.parse_data` for the "data" attribute parsing and the moving of parsing the file options to `config.load_file_options`.