Renaming (TCP/HTTP)Witnesser to (TCP/HTTP)Messengers because they are used for communicating with anyone listening.
Refactored Postman to account for all role types and scheme types when sending messages, including a shortcut if the sender is also a mailbox, then it just stores the message in the local mailbox for the recipient.

Postman prioritizes roles as Controller, Agent, Mailbox, Witness... schemes as HTTP, TCP then randomly selects if there are more than one for a role/scheme tuple.

Signed-off-by: pfeairheller <pfeairheller@gmail.com>