Fix to remove hardcoded salts as default behavior.
Removed hardcoded salts as default behavior of openHby and openHab.

Also fixed a bug that hadn't been caught yet wherein whatever was passed
to openHby was always used and a default salt was never created.
Adjusted to match the intention of the code and no tests seemed to break
from it so it must have been getting caught further down the stack.