Improvements to kli commands `mailbox add` and `witness authenticate` to work in all cases (like delegated AIDs).
New `kli aid` command for printing the AID of an alias (useful in scripts).

Added `--delpre` as a command line argument to `kli icncept` (useful in scripts).

Fix for witness publisher to account for multiple events in one cue'ed message.

Added support for witness auth codes for delegated AIDs.

Removed logging messages in the INFO case for escrow reprocessing.

Signed-off-by: pfeairheller <pfeairheller@gmail.com>