2 fixes for delegation:
- Update clone delegation to not raise an exception if the delegator's KEL is not in local kevers
- Fix event handling to walk the KEL to find an anchoring event if it exists.

These changes close #842 by allowing for a witness to return a KEL for a delegated AID without the AES couple for approved events and without the delegator's KEL.

This is the expected behavior for a witness as it should not be assumed that a witness perform watcher functionality.  Any validator will have to be introduced to the delegator or OOBI with the delegator before validating the KEL of the delegate.

Signed-off-by: pfeairheller <pfeairheller@gmail.com>