fix: allow missing "rp" field in exn message from pre 1.1.25 KERIpy
Older versions of KERIpy do not have the "rp" field in their Exchange (exn) message body at the top level. As of version 1.1.25, about June 12th 2024, the "rp" message was added for message routing, which was a backwards incompatible change since the parser was not updated to support a new protocol version for the new exn message version. This PR is a temporary workaround for this incompatibility. The plan is to tell people to move off of 1.1.x versions as soon as possible.